backend/requirements.txt
Django>=4.2,<5.0
djangorestframework<3.15
djangorestframework-simplejwt<5.3
django-cors-headers
mongoengine
pymongo
joblib
scikit-learn
pandas
numpy
channels>=4.0
channels-redis
python-dotenv
Pillow
bcrypt
